+++ src/Modal/index.tsx
@@ -1,8 +1,8 @@
 'use client'
 import * as focusTrap from 'focus-trap';
 import React, { Fragment, ElementType, HTMLProps } from 'react';
-import asModal from '../asModal';
-import { IModalContext } from '../ModalProvider/context';
+import asModal from '../asModal/index.js';
+import { IModalContext } from '../ModalProvider/context.js';
 
 export type ModalPropsWithContext = ModalProps & {
   modal?: IModalContext
