+++ src/ModalProvider/index.tsx
@@ -9,9 +9,9 @@ import React, {
 } from 'react';
 import queryString from 'qs';
 import { disableBodyScroll, enableBodyScroll } from 'body-scroll-lock';
-import generateCSS from './generateCSS';
-import ModalContext, { ModalState } from './context';
-import reducer from './reducer';
+import generateCSS from './generateCSS.js';
+import ModalContext, { ModalState } from './context.js';
+import reducer from './reducer.js';
 
 export type ModalProviderProps = {
   generateCSS?: boolean
