+++ src/index.ts
@@ -1,7 +1,7 @@
-export { default as asModal } from './asModal';
-export { default as Modal } from './Modal';
-export { default as ModalContainer } from './ModalContainer';
-export { default as ModalContext } from './ModalProvider/context';
-export { default as ModalProvider } from './ModalProvider';
-export { default as ModalToggler } from './ModalToggler';
-export { default as useModal } from './useModal';
+export { default as asModal } from './asModal/index.js';
+export { default as Modal } from './Modal/index.js';
+export { default as ModalContainer } from './ModalContainer/index.js';
+export { default as ModalContext } from './ModalProvider/context.js';
+export { default as ModalProvider } from './ModalProvider/index.js';
+export { default as ModalToggler } from './ModalToggler/index.js';
+export { default as useModal } from './useModal/index.js';
