+++ src/useModal/index.tsx
@@ -1,6 +1,6 @@
 'use client'
 import { useContext } from 'react';
-import ModalContext, { IModalContext } from '../ModalProvider/context';
+import ModalContext, { IModalContext } from '../ModalProvider/context.js';
 
 const useModal = (): IModalContext => useContext(ModalContext);
