+++ tsconfig.demo.json
@@ -2,7 +2,7 @@
   "compilerOptions": {
     "allowJs": true,
     "target": "es5",
-    "module": "commonjs",
+    "module": "NodeNext",
     "strict": true,
     "sourceMap": true,
     "jsx": "react",
