+++ tsconfig.json
@@ -3,7 +3,8 @@
     "outDir": "./dist",
     "allowJs": true,
     "target": "es5",
-    "module": "commonjs",
+    "module": "NodeNext",
+    "moduleResolution": "NodeNext",
     "strict": true,
     "sourceMap": true,
     "jsx": "react",
